# docs/api/openapi.yaml
openapi: 3.0.0
info:
  title: 'Tijori API'
  description: 'Tijori family networking platform API — phone auth, family graph, encrypted chat & media. Modules are grouped by tag in Swagger UI.'
  version: 1.0.0
servers:
  - url: /api/v1
    description: 'API v1'
paths:
...
  /admin/users:
    get:
      tags:
        - Admin
      summary: 'List users'
      operationId: adminUsersList
      parameters:
        -
          name: search
          in: query
          schema:
            type: string
        -
          name: include_trashed
          in: query
          schema:
            type: boolean
        -
          name: page
          in: query
          schema:
            type: integer
        -
          name: per_page
          in: query
          schema:
            type: integer
            maximum: 50
      responses:
        200:
          description: 'Paginated users'
      security:
        -
          bearerAuth: []
  '/admin/users/{uuid}':
    get:
      tags:
        - Admin
      summary: 'Get user detail'
      operationId: adminUsersShow
      parameters:
        -
          name: uuid
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        200:
          description: 'User detail'
      security:
        -
          bearerAuth: []
    delete:
      tags:
        - Admin
      summary: 'Ban user (soft delete)'
      operationId: adminUsersBan
      parameters:
        -
          name: uuid
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        200:
          description: 'User banned'
      security:
        -
          bearerAuth: []
    patch:
      tags:
        - Admin
      summary: 'Update user'
      operationId: adminUsersUpdate
      parameters:
        -
          name: uuid
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        200:
          description: 'Updated user'
      security:
        -
          bearerAuth: []
  '/admin/users/{uuid}/restore':
    post:
      tags:
        - Admin
      summary: 'Restore banned user'
      operationId: adminUsersRestore
      parameters:
        -
          name: uuid
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        200:
          description: 'User restored'
      security:
        -
          bearerAuth: []
  '/admin/users/{uuid}/roles':
    post:
      tags:
        - Admin
      summary: 'Assign role to user'
      operationId: adminUsersAssignRole
      parameters:
        -
          name: uuid
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        200:
          description: 'Role assigned'
      security:
        -
          bearerAuth: []
  '/admin/users/{uuid}/roles/{role}':
    delete:
      tags:
        - Admin
      summary: 'Remove role from user'
      operationId: adminUsersRemoveRole
      parameters:
        -
          name: uuid
          in: path
          required: true
          schema:
            type: string
            format: uuid
        -
          name: role
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: 'Role removed'
      security:
        -
          bearerAuth: []
...
  '/groups/{uuid}/messages':
    get:
      tags:
        - Chat
      summary: 'List encrypted messages'
      operationId: groupMessagesList
      parameters:
        -
          name: uuid
          in: path
          required: true
          schema:
            type: string
            format: uuid
        -
          name: cursor
          in: query
          schema:
            type: string
            format: uuid
        -
          name: limit
          in: query
          schema:
            type: integer
            default: 30
      responses:
        200:
          description: 'Message page'
      security:
        -
          bearerAuth: []
    post:
      tags:
        - Chat
      summary: 'Send encrypted message'
      description: 'Broadcasts message.sent on private-group.{uuid} via Reverb.'
      operationId: groupMessagesSend
      parameters:
        -
          name: uuid
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        201:
          description: 'Message sent'
        422:
          description: 'Validation error'
      security:
        -
          bearerAuth: []
...
  /admin/storage/plans:
    get:
      tags:
        - Admin
      summary: 'List all storage plans'
      operationId: adminStoragePlansList
      responses:
        200:
          description: Plans
      security:
        -
          bearerAuth: []
    post:
      tags:
        - Admin
      summary: 'Create storage plan'
      operationId: adminStoragePlansCreate
      responses:
        201:
          description: Created
      security:
        -
          bearerAuth: []
  '/admin/storage/plans/{uuid}':
    patch:
      tags:
        - Admin
      summary: 'Update storage plan'
      operationId: adminStoragePlansUpdate
      parameters:
        -
          name: uuid
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        200:
          description: Updated
      security:
        -
          bearerAuth: []
  '/admin/storage/users/{userUuid}/assignment':
    get:
      tags:
        - Admin
      summary: 'Get user plan assignment'
      operationId: adminStorageUserAssignment
      parameters:
        -
          name: userUuid
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        200:
          description: Assignment
      security:
        -
          bearerAuth: []
  '/admin/storage/users/{userUuid}/assign':
    post:
      tags:
        - Admin
      summary: 'Assign plan to user'
      operationId: adminStorageAssignPlan
      parameters:
        -
          name: userUuid
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        200:
          description: Assigned
      security:
        -
          bearerAuth: []
  '/admin/storage/assignments/{id}/revoke':
    post:
      tags:
        - Admin
      summary: 'Revoke plan assignment'
      operationId: adminStorageRevokeAssignment
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        200:
          description: Revoked
      security:
        -
          bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      description: 'Use token from POST /auth/login or POST /auth/register'
      bearerFormat: Sanctum
      scheme: bearer
...